Update README.md
To fetch Business Partners, for each of the subscribers from their respective SAP S/4HANA system > I don't understand this sentence.
As we have declared `destination` and `connectivity service` we are able to use the same service keys in combination with the tenant specific subdomain to generate the oauth token. > I don't understand this sentence either.