Update README.md
The names in the structure here must be the same as the titles of the respective pages. The only difference is that the titles in the pages should be capitalized, and here we should have them as a sentence. 

Currently, some of the entries here lead to pages with different titles. This must not happen.